backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.3
supabase==2.10.0
python-jose==3.3.0
python-multipart==0.0.6
youtube-transcript-api==0.6.3
pyjwt==2.10.1
